requirements.txt
numpy
pandas
torch
scikit-learn
matplotlib
tqdm
fastapi
uvicorn
requests
pydantic
python-multipart
sse-starlette
py-cpuinfo
